backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Processing
docx-mailmerge==0.5.0
PyPDF2==3.0.1
python-barcode==0.14.0
pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2
python-multipart==0.0.6

# Utilities
loguru==0.7.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
requests==2.31.0
websockets==12.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1